+++ packages/basehub/src/bin/main.ts
@@ -27,6 +27,8 @@ export const main = async (
   args: Args,
   opts: { forceDraft?: boolean; version: string }
 ) => {
+  const sdkBuildId = "bshb_sdk_" + Math.random().toString(16).slice(2);
+
   async function generateSDK(silent: boolean, previousSchemaHash: string) {
     logIfNotSilent(silent, "🪄 Generating...");
 
@@ -205,7 +207,7 @@ R extends Omit<MutationGenqlSelection, "transaction" | "transactionAwaitable"> &
     }
 
     // 3. append our basehub function to the end of the file.
-    const basehubExport = getBaseHubExport(draft);
+    const basehubExport = getBaseHubExport({ noStore: draft, sdkBuildId });
     if (!schemaFileContents.includes(basehubExport)) {
       schemaFileContents = schemaFileContents.concat(`\n${basehubExport}`);
     }
@@ -305,19 +307,37 @@ R extends Omit<MutationGenqlSelection, "transaction" | "transactionAwaitable"> &
       // alias react-rich-text and other packages to the generated client for better import experience
       [
         "react-rich-text",
+        "react-code-block/index",
+        "react-code-block/client",
         "api-transaction",
         "react-search",
         "analytics",
         "search",
         "next-image",
       ].map((pathsToAlias) => {
+        // ensure the directory exists
+        fs.mkdirSync(
+          path.join(basehubOutputPath, ...pathsToAlias.split("/").slice(0, -1)),
+          { recursive: true }
+        );
+
         // create a file in the output directory that aliases the package to the generated client
         fs.writeFileSync(
-          path.join(basehubOutputPath, `${pathsToAlias}.d.ts`),
+          path.join(
+            basehubOutputPath,
+            ...pathsToAlias
+              .split("/")
+              .map((p, i, { length }) => (i + 1 === length ? `${p}.d.ts` : p))
+          ),
           `export * from "basehub/${pathsToAlias}";`
         );
         fs.writeFileSync(
-          path.join(basehubOutputPath, `${pathsToAlias}.js`),
+          path.join(
+            basehubOutputPath,
+            ...pathsToAlias
+              .split("/")
+              .map((p, i, { length }) => (i + 1 === length ? `${p}.js` : p))
+          ),
           `export * from "basehub/${pathsToAlias}";`
         );
       });
@@ -498,9 +518,77 @@ process.on("unhandledRejection", (reason, promise) => {
   process.exit(1);
 });
 
-const getBaseHubExport = (noStore: boolean) => `
+const getBaseHubExport = ({
+  noStore,
+  sdkBuildId,
+}: {
+  noStore: boolean;
+  sdkBuildId: string;
+}) => `
 export type * from "@basehub/mutation-api-helpers";
 import { createFetcher } from "./runtime";
+
+export const sdkBuildId = "${sdkBuildId}";
+
+/**
+ * Returns a hash code from an object
+ * @param  {Object} obj The object to hash.
+ * @return {String}    A hash string
+ */
+function hashObject(obj: Record<string, unknown>): string {
+    const sortObjectKeys = <O extends Record<string, unknown>>(obj: O): O => {
+        if (!isObjectAsWeCommonlyCallIt(obj)) return obj;
+        return Object.keys(obj)
+            .sort()
+            .reduce((acc, key) => {
+                acc[key as keyof O] = obj[key as keyof O];
+                return acc;
+            }, {} as O);
+    };
+
+    const recursiveSortObjectKeys = <O extends Record<string, unknown>>(obj: O): O => {
+        const sortedObj = sortObjectKeys(obj);
+
+        if (!isObjectAsWeCommonlyCallIt(sortedObj)) return sortedObj;
+
+        Object.keys(sortedObj).forEach((key) => {
+            if (isObjectAsWeCommonlyCallIt(sortedObj[key as keyof O])) {
+                sortedObj[key as keyof O] = recursiveSortObjectKeys(
+                    sortedObj[key as keyof O] as O
+                ) as O[keyof O];
+            } else if (Array.isArray(sortedObj[key as keyof O])) {
+                sortedObj[key as keyof O] = (sortedObj[key as keyof O] as unknown[]).map(
+                    (item) => {
+                        if (isObjectAsWeCommonlyCallIt(item)) {
+                            return recursiveSortObjectKeys(item);
+                        } else {
+                            return item;
+                        }
+                    }
+                ) as O[keyof O];
+            }
+        });
+
+        return sortedObj;
+    };
+
+    const isObjectAsWeCommonlyCallIt = (
+        obj: unknown
+    ): obj is Record<string, unknown> => {
+        return Object.prototype.toString.call(obj) === '[object Object]';
+    };
+
+    const sortedObj = recursiveSortObjectKeys(obj);
+    const str = JSON.stringify(sortedObj);
+
+    let hash = 0;
+    for (let i = 0, len = str.length; i < len; i++) {
+        let chr = str.charCodeAt(i);
+        hash = (hash << 5) - hash + chr;
+        hash |= 0; // Convert to 32bit integer
+    }
+    return Math.abs(hash).toString();
+}
 
 // we limit options to only the ones we want to expose.
 type Options = Omit<ClientOptions, 'url' | 'method' | 'batch' | 'credentials' | 'fetch' | 'fetcher' | 'headers' | 'integrity' | 'keepalive' | 'mode' | 'redirect' | 'referrer' | 'referrerPolicy' | 'window'> & { draft?: boolean, token?: string }
@@ -523,6 +611,24 @@ type Options = Omit<ClientOptions, 'url' | 'method' | 'batch' | 'credentials' | 
  */
 export const basehub = (options?: Options) => {
   const { url, headers } = getStuffFromEnv(options);
+
+  if (!options) {
+    options = {};
+  }
+
+  options.getExtraFetchOptions = (op, body) => {
+    if (op !== 'query') return {}
+    const queryHash = hashObject(body)
+    const cacheTag = 'basehub-' + queryHash
+
+    // don't override if we're in draft mode
+    if (${noStore}) return {}
+
+    // don't override if revalidation is already being handled by the user
+    if (typeof options?.next !== 'undefined') return {}
+
+    return { next: { tags: [cacheTag] }, headers: { 'x-basehub-sdk-build-id': "${sdkBuildId}", 'x-basehub-cache-tag': cacheTag }}
+  }
 
   return {
     ...createClient(${