+++ packages/basehub/src/react/code-block/types.ts
@@ -1,0 +1,10 @@
+import { HighlighterProps } from "./highlighter";
+
+export type Snippet = {
+  id: string;
+  label?: string;
+  code: string;
+  lang: HighlighterProps["lang"];
+};
+
+export type ClientSnippet = Omit<Snippet, "code">;